Consolidate signer management components
